distributions/aws/test/e2e/requirements.txt
absl-py==0.11.0
attrs==21.2.0
boto3==1.18.50
botocore==1.21.65
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==2.0.0
Deprecated==1.2.13
docstring-parser==0.13
execnet==1.9.0
fire==0.4.0
google-api-core==1.31.4
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
iniconfig==1.1.1
jmespath==0.10.0
jsonschema==3.2.0
kfp==1.8.9
kfp-pipeline-spec==0.1.13
kfp-server-api==1.7.1
kubernetes==12.0.1
mysql-connector-python==8.0.27
oauthlib==3.1.1
packaging==21.3
pick==1.2.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==2.2.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.8
s3transfer==0.5.0
six==1.16.0
strip-hints==0.1.10
tabulate==0.8.9
termcolor==1.1.0
toml==0.10.2
typer==0.4.0
typing-extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.7
websocket-client==1.2.1
wrapt==1.13.3
